sgr-agent-erc-gemini3/requirements.txt
--extra-index-url https://erc.timetoact-group.at/
erc3>=1.1.4
google-generativeai
